+++ app/src/main/java/com/example/meditationbiorefactoring/feature_bio/presentation/measurement/brpm/BrpmEvent.kt
@@ -3,6 +3,6 @@ package com.example.meditationbiorefactoring.feature_bio.presentation.measuremen
 sealed class BrpmEvent {
     data object Start : BrpmEvent()
     data object Reset : BrpmEvent()
-    data class DataCaptured (val z: Float) : BrpmEvent()
+    data class DataCaptured (val z: Double) : BrpmEvent()
     data object NavigateClick : BrpmEvent()
 }
